refactor: move mixins to publishable_entity.py
This is to prevent circular dependencies.
