+++ openedx_learning/apps/authoring/publishing/models/container.py
@@ -4,7 +4,7 @@ Container and ContainerVersion models
 from django.core.exceptions import ValidationError
 from django.db import models
 
-from ..model_mixins import PublishableEntityMixin, PublishableEntityVersionMixin
+from .publishable_entity import PublishableEntityMixin, PublishableEntityVersionMixin
 from .entity_list import EntityList
